improve sprites generation
