+++ .gitignore
@@ -1,1 +1,2 @@
 *.bas
+__pycache__
